Patch 6.2.359 (after 6.2.352) Problem:    Compiler warning for long to int type cast. Solution:   Add explicit type cast. Files:	    src/ex_cmds.c

+++ src/ex_cmds.c
@@ -5779,7 +5779,7 @@ ex_sign(eap)
 	    else if (STRNCMP(arg, "buffer=", 7) == 0)
 	    {
 		arg += 7;
-		buf = buflist_findnr(getdigits(&arg));
+		buf = buflist_findnr((int)getdigits(&arg));
 		if (*skipwhite(arg) != NUL)
 		    EMSG(_(e_trailing));
 		break;
